+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 ## v1.0.0
 - Initial release
+- add `splines.SplineVessels`, a thin wrapper to `synthsplines`.